refactor app.py (0/6 prep): promote post_to_omg_lol to helpers/omg_lol.py
post_to_omg_lol is called from two future blueprints (tasks via
/tasks/add and cockpit via /publish), so it can't be cockpit-internal
as the original audit assumed. Promoted to a shared helper now so
sub-phase 1 (tasks) can import it cleanly.

OMG_LOL_API_KEY and OMG_LOL_ADDRESS continue to be read from
os.environ at call time. No behavior change.
